Update club taxes payload handling
