+++ meClub/src/screens/dashboard/ConfiguracionScreen.jsx
@@ -175,22 +175,6 @@ const sanitizeServicesForPayload = (services) => {
     .filter((value) => value !== null);
 };
 
-const sanitizeTaxesForPayload = (taxes) => {
-  if (!Array.isArray(taxes)) return [];
-  return taxes.map((tax) => {
-    const porcentaje = tax?.porcentaje ?? '';
-    const normalized = Number(String(porcentaje).replace(',', '.'));
-    const payload = {
-      nombre: tax?.nombre ?? '',
-      porcentaje: Number.isFinite(normalized) ? normalized : 0,
-    };
-    if (tax?.id && !String(tax.id).startsWith('tmp-')) {
-      payload.id = tax.id;
-    }
-    return payload;
-  });
-};
-
 const initialSaveStatus = Object.keys(STATUS_LABELS).reduce((acc, key) => {
   acc[key] = { state: 'idle', message: '' };
   return acc;
@@ -619,7 +603,7 @@ export default function ConfiguracionScreen({ go }) {
 
     operations.push(
       runOperation('taxes', async () => {
-        const response = await updateClubTaxes(sanitizeTaxesForPayload(form.impuestos));
+        const response = await updateClubTaxes(form.impuestos);
         setForm((prev) => ({
           ...prev,
           impuestos: normalizeTaxes(response),
